Abuse QML State System
The animations previously were designed to be played once, then the face
returns to idle.  However if we want to use logic like (if we are going
forward make the robot look happy, we can only make him look happy for a
short amount of time before he returns to his "idle" state (where we can
replay again, but not idea).  So instead I've attempted to abuse the
state-transition system to play the animations until returning to idle
state.  So now the character can be happy until he's not.  The
limitation is that no matter what you have to return to idle state
before changing to a new state.

So this could just be a terrible Idea (and is also a poor example of
qml), but I think it opens up some possibilties better demo behavior
using the current animations.  But if not we can always revert.